refactor: move copyLinkData
